refactors UI elements and ViewModels
